+++ index.html
@@ -351,6 +351,18 @@
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="utils.js"></script>
+    <script src="modals.js"></script>
+    <script src="dragAndDrop.js"></script>
+    <script src="associates.js"></script>
+    <script src="render.js"></script>
+    <script src="newButton.js"></script>
+    <script src="clearDataButton.js"></script>
+    <script src="headcountButton.js"></script>
+    <script src="presetsButton.js"></script>
+    <script src="startSessionButton.js"></script>
+    <script src="eventListeners.js"></script>
+    <script src="areaEventListeners.js"></script>
     <script src="main.js"></script>
     <script>
         // Initialize tooltips
